Squeeze all function in one file
